+++ spec/rdkafka/producer/delivery_report_spec.rb
@@ -1,7 +1,7 @@
 require "spec_helper"
 
 describe Rdkafka::Producer::DeliveryReport do
-  subject { Rdkafka::Producer::DeliveryReport.new(2, 100, "error") }
+  subject { Rdkafka::Producer::DeliveryReport.new(2, 100, "topic", "error") }
 
   it "should get the partition" do
     expect(subject.partition).to eq 2
@@ -9,6 +9,10 @@ describe Rdkafka::Producer::DeliveryReport do
 
   it "should get the offset" do
     expect(subject.offset).to eq 100
+  end
+
+  it "should get the topic_name" do
+    expect(subject.topic_name).to eq "topic"
   end
 
   it "should get the error" do